requirements.txt
torch
torchvision
fastapi
uvicorn
python-multipart
pillow
matplotlib
seaborn
scikit-learn
tqdm
numpy
pandas
